# sdk/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/cortex/agents/{agent}": {
            "get": {
                "operationId": "cortex.agents.show",
                "tags": [
                    "Agent"
                ],
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "required": true,
                        "description": "The agent slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AgentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AgentResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "patch": {
                "operationId": "cortex.agents.update",
                "tags": [
                    "Agent"
                ],
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "required": true,
                        "description": "The agent slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAgentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AgentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AgentResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "cortex.agents.destroy",
                "tags": [
                    "Agent"
                ],
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "required": true,
                        "description": "The agent slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/cortex/servers/{server}/instructions": {
            "get": {
                "operationId": "cortex.servers.instructions.show",
                "tags": [
                    "McpInstruction"
                ],
                "parameters": [
                    {
                        "name": "server",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`McpInstructionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/McpInstructionResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "cortex.servers.instructions.destroy",
                "tags": [
                    "McpInstruction"
                ],
                "parameters": [
                    {
                        "name": "server",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/cortex/servers/{server}/instructions/versions/{version}/publish": {
            "post": {
                "operationId": "cortex.servers.instructions.versions.publish",
                "tags": [
                    "McpInstruction"
                ],
                "parameters": [
                    {
                        "name": "server",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`McpInstructionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/McpInstructionResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/cortex/prompts/{prompt}": {
            "get": {
                "operationId": "cortex.prompts.show",
                "tags": [
                    "Prompt"
                ],
                "parameters": [
                    {
                        "name": "prompt",
                        "in": "path",
                        "required": true,
                        "description": "The prompt slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PromptResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PromptResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "patch": {
                "operationId": "cortex.prompts.update",
                "tags": [
                    "Prompt"
                ],
                "parameters": [
                    {
                        "name": "prompt",
                        "in": "path",
                        "required": true,
                        "description": "The prompt slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePromptRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`PromptResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PromptResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "cortex.prompts.destroy",
                "tags": [
                    "Prompt"
                ],
                "parameters": [
                    {
                        "name": "prompt",
                        "in": "path",
                        "required": true,
                        "description": "The prompt slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/cortex/prompts/{prompt}/versions/{version}/publish": {
            "post": {
                "operationId": "cortex.prompts.versions.publish",
                "tags": [
                    "PromptVersion"
                ],
                "parameters": [
                    {
                        "name": "prompt",
                        "in": "path",
                        "required": true,
                        "description": "The prompt slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`PromptResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PromptResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/cortex/tools/{tool}/description": {
            "get": {
                "operationId": "cortex.tools.description.show",
                "tags": [
                    "ToolDescription"
                ],
                "parameters": [
                    {
                        "name": "tool",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ToolDescriptionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ToolDescriptionResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "cortex.tools.description.destroy",
                "tags": [
                    "ToolDescription"
                ],
                "parameters": [
                    {
                        "name": "tool",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/cortex/tools/{tool}/description/versions/{version}/publish": {
            "post": {
                "operationId": "cortex.tools.description.versions.publish",
                "tags": [
                    "ToolDescription"
                ],
                "parameters": [
                    {
                        "name": "tool",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ToolDescriptionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ToolDescriptionResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AgentResource": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "provider": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "model": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "settings": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "additionalProperties": {}
                    },
                    "tools": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "prompt": {
                        "type": "string"
                    },
                    "prompt_version": {
                        "type": "integer"
                    },
                    "sub_agents": {
                        "type": "array",
                        "items": {}
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "name",
                    "slug",
                    "description",
                    "provider",
                    "model",
                    "settings",
                    "tools",
                    "created_at",
                    "updated_at"
                ],
                "title": "AgentResource"
            },
...
            "McpInstructionResource": {
                "type": "object",
                "properties": {
                    "server": {
                        "type": "string"
                    },
                    "published_version": {
                        "type": "integer"
                    },
                    "published_content": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "server",
                    "published_version",
                    "published_content",
                    "created_at",
                    "updated_at"
                ],
                "title": "McpInstructionResource"
            },
...
            "PromptResource": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "published_version": {
                        "$ref": "#/components/schemas/PromptVersionResource"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "name",
                    "slug",
                    "description",
                    "created_at",
                    "updated_at"
                ],
                "title": "PromptResource"
            },
            "PromptVersionResource": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "integer"
                    },
                    "content": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "version",
                    "content",
                    "created_at"
                ],
                "title": "PromptVersionResource"
            },
...
            "ToolDescriptionResource": {
                "type": "object",
                "properties": {
                    "tool": {
                        "type": "string"
                    },
                    "published_version": {
                        "type": "integer"
                    },
                    "published_content": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "tool",
                    "published_version",
                    "published_content",
                    "created_at",
                    "updated_at"
                ],
                "title": "ToolDescriptionResource"
            },
...
            "UpdateAgentRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "provider": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "model": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "settings": {
                        "type": "object",
                        "properties": {
                            "temperature": {
                                "type": "number",
                                "minimum": 0,
                                "maximum": 2
                            },
                            "max_steps": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "max_tokens": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "top_p": {
                                "type": "number",
                                "minimum": 0,
                                "maximum": 1
                            }
                        },
                        "required": [
                            "temperature",
                            "max_steps",
                            "max_tokens",
                            "top_p"
                        ]
                    },
                    "tools": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                ""
                            ]
                        }
                    },
                    "prompt": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "prompt_version": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 1
                    },
                    "sub_agents": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "uniqueItems": true
                    }
                },
                "title": "UpdateAgentRequest"
            },
            "UpdatePromptRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "title": "UpdatePromptRequest"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}